Adding clickable views setting
Setting the clickable views of a Facebook native ad to be the child
views of the view (icon, main image, title, text etc.) instead of the
view.
We do this to avoid the white space in the view container for the
native ad unit to be clickable which leads to accidental clicks.